Fix Travis errors
