+++ high_dimensional_sampling/optimisation/__init__.py
@@ -1,2 +1,5 @@
 from .randomoptimisation import RandomOptimisation  # noqa: F401
-from .randomoptimisation import HdsPsInterface  # noqa: F401
+
+""" The `pyscannerbit` implementation is excluded from unit tests, as we 
+-- unfortunately -- did not get pyscannerbit to work within Travis """
+from .pyscannerbit import PyScannerBit  # noqa: F401
